+++ aladdin/lib/git.py
@@ -1,6 +1,13 @@
+import logging
 import os
 import subprocess
+import tempfile
+import sys
+from contextlib import contextmanager, suppress
+
 from aladdin.lib.utils import working_directory
+from aladdin.lib.helm_rules import HelmRules
+from aladdin.config import load_git_configs
 
 
 class Git:
@@ -87,3 +94,44 @@ class Git:
             return output[0] if output else None
         except subprocess.CalledProcessError:
             return None
+
+
+@contextmanager
+def clone_and_checkout(githash, repo_name=None):
+    current_hash = None
+    current_repo = None
+    with suppress(subprocess.CalledProcessError):
+        current_repo = Git.get_repo_name()
+        current_hash = Git.get_hash()
+
+    if (
+        current_hash and
+        current_repo and
+        (current_repo == repo_name or not repo_name) and
+        current_hash == githash and
+        (Git.clean_working_tree() or HelmRules.debug)
+    ):
+        yield Git.get_base_directory()
+        return
+
+    if not repo_name:
+        logging.error(f"No repo found or specified")
+        return sys.exit(1)
+
+    git_account = load_git_configs()["account"]
+    git_url = f"git@github.com:{git_account}/{repo_name}.git"
+
+    with tempfile.TemporaryDirectory() as tmpdirname:
+        try:
+            Git.clone(git_url, tmpdirname)
+        except subprocess.CalledProcessError:
+            logging.warn(f"Could not clone repo {git_url}. Does it exist?")
+            return sys.exit(1)
+        try:
+            Git.checkout(tmpdirname, githash)
+        except subprocess.CalledProcessError:
+            logging.warn(
+                f"Could not checkout to ref '{githash}' in repo {git_url}. Have you pushed it to remote?"
+            )
+            return sys.exit(1)
+        yield tmpdirname
